+++ 00-starter/src/main/java/com/blog/demo/dto/UserResponse.java
@@ -1,7 +1,6 @@
 package com.blog.demo.dto;
 
 import com.blog.demo.entity.User;
-import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class UserResponse {
 
-    private int id;
+    private Long id;
     private String username;
 
     public UserResponse(User user) {
