+++ 00-starter/src/main/java/com/blog/demo/entity/FollowerID.java
@@ -2,7 +2,6 @@ package com.blog.demo.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,8 +13,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class FollowerID implements Serializable {
     @Column(name = "user_id", nullable = false)
-    private int receiverId;
+    private Long receiverId;
 
     @Column(name = "follower_id", nullable = false)
-    private int actorId;
+    private Long actorId;
 }
